Add executive summary display in workplan-run action (#546)
* Add utilities for displaying colorful console output
* add try_deserialize for uninterruped iterative deserialization
* Relocate safe_name property to process handle base class
* Enable on-demand run-directory resolution
* Fix use of deprecated typing.Mapping
* Add status, launcher/run "foreign keys" to process handle
* Add sentinel paths and state_dir to workplan run
* Refactor state mgmt into state repo, track new state fields during run
* Add and use executive summary models in place of printing
* Add "further actions" prompt printing
* do not use global run repo, rename
* add autouse orch fixture for clean output, remove now-useless fixtures
* simplify sentinel addition / remove attach_sentinel
* Avoid storing task status in two locations
* Extract callbacks from launcher and move into orchestrator
* Adjust tests previously writing blindly into local directory
* Replace "unused" test params with usefixtures for linter
* return run summary from dag runner for display by CLI
* log full exception details
* avoid updating symlink with same value
* improve model schema documentation
* Fix proxy file name bug
* Fix null-ref-get-dag-status-on-local-launcher bug
* Serialize working_dir on live step
* Fix incorrect-attribute-name-in-override bug
* Fix test regression bug (workplan name not in run output)

+++ cstar/orchestration/tracking.py
@@ -1,5 +1,6 @@
 import asyncio
 import typing as t
+from collections.abc import Mapping
 from datetime import datetime
 from pathlib import Path
 
@@ -36,8 +37,11 @@ class WorkplanRun(BaseModel):
     environment: dict[str, str] = Field(default_factory=dict)
     """The environment variables at the time of the run."""
 
-    user_variables: t.Mapping[str, str] = Field(default_factory=dict)
+    user_variables: Mapping[str, str] = Field(default_factory=dict)
     """User-supplied runtime variable overrides."""
+
+    sentinels: set[Path] = Field(default_factory=set[Path])
+    """State files expected to be created during execution of the run."""
 
     @staticmethod
     def get_default_run_id(uri: str) -> str:
@@ -56,6 +60,16 @@ class WorkplanRun(BaseModel):
             wp = deserialize(local_path, Workplan)
 
         return slugify(wp.name)
+
+    @property
+    def state_dir(self) -> Path:
+        """The path to the directory containing state files for the run.
+
+        Returns
+        -------
+        Path
+        """
+        return StateDirectoryManager.run_state_dir(run_id=self.run_id)
 
 
 class TrackingRepository(LoggingMixin):
@@ -213,6 +227,10 @@ class TrackingRepository(LoggingMixin):
         WorkplanRun | None
             The record when it can be located in history or latest runs, otherwise `None`.
         """
+        if not run_id:
+            msg = "A valid run-id was not provided; unable to retrieve run"
+            raise ValueError(msg)
+
         run_path = self._find_run_path(run_id, run_date)
 
         if not run_path.exists():
@@ -247,8 +265,9 @@ class TrackingRepository(LoggingMixin):
         if not latest_path.parent.exists():
             latest_path.parent.mkdir(parents=True)
 
-        await asyncio.to_thread(latest_path.unlink, missing_ok=True)
-        await asyncio.to_thread(latest_path.symlink_to, run_path)
+        if latest_path.resolve() != run_path:
+            await asyncio.to_thread(latest_path.unlink, missing_ok=True)
+            await asyncio.to_thread(latest_path.symlink_to, run_path)
 
         msg = f"Run persisted to: {run_path}"
         self.log.debug(msg)
